chore(lib/cctp): reorg integration test utils
