+++ lib/cctp/devclient.go
@@ -3,20 +3,29 @@ package cctp
 import (
 	"context"
 	"crypto/ecdsa"
+	"math/big"
 	"slices"
 	"sync"
+	"testing"
 
+	"github.com/omni-network/omni/lib/bi"
+	"github.com/omni-network/omni/lib/cctp/testutil"
 	"github.com/omni-network/omni/lib/errors"
 	"github.com/omni-network/omni/lib/ethclient"
 	"github.com/omni-network/omni/lib/evmchain"
 	"github.com/omni-network/omni/lib/expbackoff"
 	"github.com/omni-network/omni/lib/log"
+	"github.com/omni-network/omni/lib/tutil"
 	"github.com/omni-network/omni/lib/xchain"
 
+	"github.com/ethereum/go-ethereum"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
+
+	"github.com/stretchr/testify/require"
 )
 
 // DevClient is a devnet implementation of the CCTP client.
@@ -217,4 +226,135 @@ func newMessageTransmitters(clients map[uint64]ethclient.Client) (map[uint64]*Me
 	}
 
 	return transmitters, addrs, nil
+}
+
+// StartTestClient is starts a dev client for integration tests.
+// It creates a new attester account, enables it on all chains, and starts attesting.
+// It requires rpcs to be anvil forks run with auto impersonation.
+// It returns the client.
+func StartTestClient(
+	ctx context.Context,
+	t *testing.T,
+	xprov xchain.Provider,
+	chains []evmchain.Metadata,
+	clients map[uint64]ethclient.Client,
+) Client {
+	t.Helper()
+
+	attesterPk, attesterAddr := testutil.NewAccount(t)
+
+	// Create CCTP client, start attesting
+	cctpClient := NewDevClient(attesterPk, clients)
+
+	// Enable attester on all chains
+	enableAttester(ctx, t, clients, attesterAddr)
+
+	// Start attesting
+	err := cctpClient.AttestForever(ctx, chains, xprov)
+	require.NoError(t, err)
+
+	return cctpClient
+}
+
+// enableAttester enables the attester on each chains MessageTransmitter, and sets the signature threshold to 1.
+func enableAttester(ctx context.Context, t *testing.T, clients map[uint64]ethclient.Client, newAttester common.Address) {
+	t.Helper()
+
+	for chainID, client := range clients {
+		mtAddr, ok := MessageTransmitterAddr(chainID)
+		require.True(t, ok)
+
+		mt, err := NewMessageTransmitter(mtAddr, client)
+		require.NoError(t, err)
+
+		// AttesterManager is account allowed to enable attesters
+		mngr, err := mt.AttesterManager(&bind.CallOpts{Context: ctx})
+		require.NoError(t, err)
+
+		// Send unsigned MessageTransmitter.enableAttester tx from attester manager
+		// This requires anvil auto impersonation
+		calldata, err := messageTransmitterABI.Pack("enableAttester", newAttester)
+		require.NoError(t, err)
+		txHash, err := sendUnsignedTransaction(ctx, client, txArgs{
+			from:  mngr,
+			to:    mtAddr,
+			value: nil,
+			data:  calldata,
+		})
+		require.NoError(t, err)
+
+		_, err = bind.WaitMinedHash(ctx, client, txHash)
+		require.NoError(t, err)
+
+		// Verify attester is enabled
+		enabled, err := mt.IsEnabledAttester(&bind.CallOpts{Context: ctx}, newAttester)
+		require.NoError(t, err)
+		require.True(t, enabled, "attester not enabled on chain %d", chainID)
+
+		log.Info(ctx, "Enabled attester", "chain", chainID, "attester", newAttester)
+
+		// Reduce signature threshold to 1 (number of attesters required)
+		calldata, err = messageTransmitterABI.Pack("setSignatureThreshold", bi.One())
+		require.NoError(t, err)
+		txHash, err = sendUnsignedTransaction(ctx, client, txArgs{
+			from:  mngr,
+			to:    mtAddr,
+			value: nil,
+			data:  calldata,
+		})
+		require.NoError(t, err)
+
+		_, err = bind.WaitMinedHash(ctx, client, txHash)
+		require.NoError(t, err)
+
+		threshold, err := mt.SignatureThreshold(&bind.CallOpts{Context: ctx})
+		require.NoError(t, err)
+		tutil.RequireEQ(t, threshold, bi.One(), "signature threshold not set to 1 on chain %d", chainID)
+
+		log.Info(ctx, "Signature threshold set", "chain", chainID, "threshold", threshold)
+	}
+}
+
+type txArgs struct {
+	from  common.Address
+	to    common.Address
+	value *big.Int
+	data  []byte
+}
+
+// This is used to send auto impersonated txs on anvil.
+func sendUnsignedTransaction(ctx context.Context, client ethclient.Client, args txArgs) (common.Hash, error) {
+	value := bi.Zero()
+	if args.value != nil {
+		value = args.value
+	}
+
+	to := args.to
+	msg := ethereum.CallMsg{
+		From:  args.from,
+		To:    &to,
+		Value: value,
+		Data:  args.data,
+	}
+
+	gas, err := client.EstimateGas(ctx, msg)
+	if err != nil {
+		return common.Hash{}, errors.Wrap(err, "estimate gas")
+	}
+
+	jsonArgs := map[string]any{
+		"from":  args.from,
+		"to":    to,
+		"value": (*hexutil.Big)(value),
+		"data":  hexutil.Bytes(args.data),
+		"gas":   hexutil.Uint64(gas),
+	}
+
+	var result common.Hash
+	err = client.CallContext(ctx, &result, "eth_sendTransaction", jsonArgs)
+	if err != nil {
+		return common.Hash{}, errors.Wrap(err, "send transaction")
+	}
+
+	return result, nil
 }
